Update unit test structure to new module structure
