+++ tests/test_s3.py
@@ -1,0 +1,10 @@
+
+
+class TestOdimFilePath:
+    # TODO - add unit tests
+
+    def test_parse_file_name(self):
+        """"""
+        pass
+        # baltrad/hdf5/fivan/2016/10/25/fivan_vp_20161025T2100Z_0x7_147742969449.h5
+        # baltrad/hdf5/fiuta/2021/11/14/fiuta_vp_20211114T214500Z_0xb.h5
